timescaledb/requirements.txt
python-dotenv
python-dateutil
requests